requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Werkzeug==2.3.7
email-validator==2.1.0
python-dotenv==1.0.0
itsdangerous==2.1.2
Flask-Mail